[frontend] new filters in Attack Patterns of a Threat TRIAL (#7287)
